pyproject.toml
# AVA Trading Platform - Modern Python Project Configuration
# Uses PEP 621 standards and uv/pip for dependency management

[project]
name = "ava-trading-platform"
version = "2.0.0"
description = "AI-Powered Trading Platform with Multi-Agent Intelligence"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "AVA Team", email = "team@ava-trading.com" }
]
keywords = ["trading", "ai", "options", "stocks", "sports-betting", "langchain"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    # Core Framework
    "fastapi>=0.122.0",
    "uvicorn[standard]>=0.38.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",

    # Async & HTTP
    "aiohttp>=3.13.2",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.3",
    "sse-starlette>=3.0.3",
    "websockets>=15.0.1",

    # Database
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.44",
    "redis>=7.1.0",

    # AI/ML Core
    "anthropic>=0.75.0",
    "openai>=2.8.1",
    "groq>=0.36.0",
    "ollama>=0.6.1",
    "instructor>=1.13.0",

    # LangChain Ecosystem
    "langchain>=0.3.25",
    "langchain-core>=0.3.61",
    "langchain-anthropic>=0.3.15",
    "langchain-openai>=0.3.18",
    "langchain-community>=0.3.25",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.4.10",
    "langgraph-checkpoint>=3.0.1",
    "langgraph-prebuilt>=1.0.5",
    "langsmith>=0.4.49",

    # Data Science
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pandas-ta>=0.4.71b0",
    "scipy>=1.16.3",
    "scikit-learn>=1.7.2",

    # ML/Deep Learning
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "sentence-transformers>=5.1.2",
    "huggingface-hub>=0.36.0",

    # Vector Database
    "chromadb>=1.1.1",

    # Finance & Trading
    "yfinance>=0.2.66",
    "robin-stocks>=3.4.0",
    "tradingview-ta>=3.3.0",

    # Options Pricing
    "py-vollib>=2.1.0",

    # Observability
    "opentelemetry-api>=1.38.0",
    "opentelemetry-sdk>=1.38.0",
    "opentelemetry-exporter-otlp>=1.38.0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "opentelemetry-instrumentation-aiohttp-client>=0.59b0",
    "opentelemetry-instrumentation-redis>=0.59b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.59b0",
    "structlog>=25.4.0",
    "prometheus-client>=0.22.1",

    # Utilities
    "orjson>=3.11.4",
    "tenacity>=9.1.2",
    "backoff>=2.2.1",
    "cachetools>=6.2.2",
    "rich>=14.2.0",
    "loguru>=0.7.3",
    "typer>=0.20.0",

    # Web Scraping & Parsing
    "beautifulsoup4>=4.14.2",
    "pdfplumber>=0.11.8",

    # Social & News
    "praw>=7.8.1",

    # Notifications
    "python-telegram-bot>=21.0",

    # Validation
    "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.0",
    "pytest-xdist>=3.5.0",
    "hypothesis>=6.130.6",
    "mypy>=1.15.0",
    "ruff>=0.11.13",
    "black>=25.1.0",
    "isort>=6.0.1",
    "pre-commit>=4.5.0",
    "bandit>=1.8.3",
]

test = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.0",
    "httpx>=0.28.1",
    "faker>=37.3.0",
    "factory-boy>=3.3.3",
]

docs = [
    "mkdocs>=1.7.0",
    "mkdocs-material>=9.7.5",
    "mkdocstrings[python]>=0.29.1",
]

[project.scripts]
ava = "src.ava.cli:app"
ava-server = "backend.main:run_server"

[project.urls]
Homepage = "https://github.com/ava-trading/ava"
Documentation = "https://docs.ava-trading.com"
Repository = "https://github.com/ava-trading/ava"

# Build System
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ava", "backend"]

# Ruff Configuration (modern Python linter/formatter)
[tool.ruff]
target-version = "py311"
line-length = 100
fix = true

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
    "ASYNC",  # flake8-async
    "S",      # flake8-bandit (security)
    "DTZ",    # flake8-datetimez
    "T20",    # flake8-print
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "S101",   # assert is fine in tests
    "T201",   # print in CLI is fine
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ARG001", "ARG002"]
"**/migrations/**/*.py" = ["ARG001"]

[tool.ruff.lint.isort]
known-first-party = ["ava", "backend"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# MyPy Configuration
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_any_generics = true
warn_redundant_casts = true
plugins = [
    "pydantic.mypy"
]

[[tool.mypy.overrides]]
module = [
    "yfinance.*",
    "robin_stocks.*",
    "pandas_ta.*",
    "chromadb.*",
    "tradingview_ta.*",
]
ignore_missing_imports = true

# Pytest Configuration
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "-p", "no:warnings",
    "--cov=src/ava",
    "--cov=backend",
    "--cov-report=term-missing",
    "--cov-report=html:coverage_html",
    "--cov-fail-under=70",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# Coverage Configuration
[tool.coverage.run]
source = ["src/ava", "backend"]
branch = true
parallel = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Black Configuration (if used alongside ruff)
[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
)/
'''

# Bandit Configuration (Security)
[tool.bandit]
exclude_dirs = ["tests", "migrations"]
skips = ["B101", "B311"]
